refactor(adapters,review): split openai/anthropic adapters into _decode + _encode halves
Addresses Copilot review on PR #230 (comments #3 and #4) — the
openai_messages.py and anthropic_messages.py modules sat at 394 and
373 lines respectively, over the ≤300-line module guideline in
AGENTS.md.

Split pattern mirrors the existing private `_messages_common.py`
convention:

  openai_messages.py     (394 → 193 lines, public surface only)
    ├── _openai_decode.py    (decoder helpers + ID prefix constants)
    └── _openai_encode.py    (encoder helpers)

  anthropic_messages.py  (373 → 146 lines, public surface only)
    ├── _anthropic_decode.py (decoder helpers + ID prefix constants)
    └── _anthropic_encode.py (encoder helpers)

Behaviour preserved bit-for-bit:
- No public API change (`from_/to_*` symbols unchanged).
- adapters/__init__.py re-exports unchanged.
- Existing 38 adapter tests pass with zero modification.
- Full suite: 1002 passed, 8 skipped.
- make ci + make scorecard-check + make llms-check all clean.

The new private modules follow the same naming pattern as
`_messages_common.py` (single-underscore prefix, not part of the
public contract). AGENTS.md module map already lists only public
adapter modules so no entry was added — same convention as
`_messages_common.py`.

`gemini_contents.py` is already at 299 lines (under the guideline)
and is left untouched.
