+++ src/contextweaver/adapters/openai_messages.py
@@ -30,12 +30,15 @@ Mapping rules:
 :func:`to_openai_messages` is the inverse of :func:`from_openai_messages`
 for any well-formed input.
 
+The decoder helpers live in :mod:`._openai_decode` and the encoder
+helpers in :mod:`._openai_encode` to keep this public surface focused
+and within the repo's ≤300-line module guideline.
+
 Issue #219 (slice of #194).
 """
 
 from __future__ import annotations
 
-import json
 import logging
 from typing import TYPE_CHECKING, Any
 
@@ -45,6 +48,18 @@ from contextweaver.adapters._messages_common import (
     ingest_into_manager,
     strip_prefix,
 )
+from contextweaver.adapters._openai_decode import (
+    _PREFIX_TOOL_RESULT,
+    _from_assistant,
+    _from_system,
+    _from_tool,
+    _from_user,
+)
+from contextweaver.adapters._openai_encode import (
+    _collect_assistant_tool_calls,
+    _restore_assistant_content,
+    _tool_call_payload,
+)
 from contextweaver.exceptions import CatalogError
 from contextweaver.types import ContextItem, ItemKind
 
@@ -52,19 +67,6 @@ if TYPE_CHECKING:
     from contextweaver.context.manager import ContextManager
 
 logger = logging.getLogger("contextweaver.adapters")
-
-# ID prefix carries enough information to round-trip the OpenAI message:
-# - "openai:tool_call:<tool_call_id>"   → the assistant's tool-call entry
-# - "openai:tool_result:<tool_call_id>" → the matching role="tool" response
-# - "openai:{user,assistant,system}:<idx>" → turn entries
-_PREFIX_TOOL_CALL = "openai:tool_call:"
-_PREFIX_TOOL_RESULT = "openai:tool_result:"
-_PREFIX_USER = "openai:user:"
-_PREFIX_ASSISTANT = "openai:assistant:"
-_PREFIX_SYSTEM = "openai:system:"
-
-
-# --- Public: from_openai_messages ---
 
 
 def from_openai_messages(
@@ -116,9 +118,6 @@ def from_openai_messages(
     ingest_into_manager(items, into)
     logger.debug("from_openai_messages: messages_in=%d, items_out=%d", len(messages), len(items))
     return items
-
-
-# --- Public: to_openai_messages ---
 
 
 def to_openai_messages(items: list[ContextItem]) -> list[dict[str, Any]]:
@@ -192,203 +191,3 @@ def to_openai_messages(items: list[ContextItem]) -> list[dict[str, Any]]:
             )
         i += 1
     return messages
-
-
-# --- Per-role decoders ---
-
-
-def _from_system(msg: dict[str, Any], idx: int) -> ContextItem:
-    return _make_turn(msg, idx, role="system", kind=ItemKind.policy, prefix=_PREFIX_SYSTEM)
-
-
-def _from_user(msg: dict[str, Any], idx: int) -> ContextItem:
-    return _make_turn(msg, idx, role="user", kind=ItemKind.user_turn, prefix=_PREFIX_USER)
-
-
-def _make_turn(
-    msg: dict[str, Any], idx: int, *, role: str, kind: ItemKind, prefix: str
-) -> ContextItem:
-    return ContextItem(
-        id=f"{prefix}{idx}",
-        kind=kind,
-        text=_string_content(msg, idx, role),
-        metadata={"role": role},
-    )
-
-
-def _from_assistant(
-    msg: dict[str, Any], idx: int, seen_tool_call_ids: set[str]
-) -> list[ContextItem]:
-    """Expand an assistant message into agent_msg + N tool_call items.
-
-    Metadata keys ``content_is_null`` and ``original_content`` preserve
-    the three input shapes (``None`` / ``""`` / list-of-parts) so the
-    inverse adapter can re-emit them exactly (PR #230 review).
-    """
-    out: list[ContextItem] = []
-    raw_content = msg.get("content")
-    text = "" if raw_content is None else _string_content(msg, idx, "assistant")
-    out.append(
-        ContextItem(
-            id=f"{_PREFIX_ASSISTANT}{idx}",
-            kind=ItemKind.agent_msg,
-            text=text,
-            metadata={
-                "role": "assistant",
-                "content_is_null": raw_content is None,
-                "original_content": raw_content,
-                "assistant_idx": idx,
-            },
-        )
-    )
-
-    tool_calls = msg.get("tool_calls") or []
-    if not isinstance(tool_calls, list):
-        raise CatalogError(f"OpenAI assistant message at index {idx} has non-list tool_calls")
-
-    for tc_idx, tc in enumerate(tool_calls):
-        loc = f"OpenAI tool_call at index [{idx}].tool_calls[{tc_idx}]"
-        expect_dict(tc, label=loc)
-        tc_id = tc.get("id")
-        if not tc_id:
-            raise CatalogError(f"{loc} missing id")
-        fn = tc.get("function") or {}
-        expect_dict(fn, label=f"OpenAI tool_call {tc_id!r} function payload")
-        fn_name = fn.get("name", "")
-        # Arguments are a JSON-encoded string in OpenAI's schema; preserve
-        # raw text + parsed args in metadata so to_openai_messages can
-        # re-emit the canonical shape.
-        args_str = fn.get("arguments", "")
-        if not isinstance(args_str, str):
-            args_str = json.dumps(args_str)
-        out.append(
-            ContextItem(
-                id=f"{_PREFIX_TOOL_CALL}{tc_id}",
-                kind=ItemKind.tool_call,
-                text=args_str,
-                metadata={
-                    "tool_call_id": tc_id,
-                    "tool_call_type": tc.get("type", "function"),
-                    "function_name": fn_name,
-                    "arguments": args_str,
-                    "assistant_idx": idx,
-                },
-            )
-        )
-        seen_tool_call_ids.add(tc_id)
-    return out
-
-
-def _from_tool(msg: dict[str, Any], idx: int, seen_tool_call_ids: set[str]) -> ContextItem:
-    tool_call_id = msg.get("tool_call_id")
-    if not tool_call_id:
-        raise CatalogError(f"OpenAI tool message at index {idx} missing tool_call_id")
-    if tool_call_id not in seen_tool_call_ids:
-        # The tool_call_id must reference a prior assistant tool_calls entry
-        # so the resulting ContextItem.parent_id points at a real item (PR
-        # #230 review). Mirrors the Gemini adapter's unmatched-response check.
-        raise CatalogError(
-            f"OpenAI tool message at index {idx} has tool_call_id={tool_call_id!r} "
-            "that does not match any prior assistant tool_calls entry"
-        )
-    content = _string_content(msg, idx, "tool")
-    return ContextItem(
-        id=f"{_PREFIX_TOOL_RESULT}{tool_call_id}",
-        kind=ItemKind.tool_result,
-        text=content,
-        metadata={"role": "tool", "tool_call_id": tool_call_id},
-        parent_id=f"{_PREFIX_TOOL_CALL}{tool_call_id}",
-    )
-
-
-# --- to_openai_messages helpers ---
-
-
-def _collect_assistant_tool_calls(items: list[ContextItem], start: int) -> list[dict[str, Any]]:
-    """Collect tool_call items that follow an assistant agent_msg.
-
-    A tool_call belongs to the immediately-preceding agent_msg iff its
-    ``metadata["assistant_idx"]`` matches the agent_msg's input index. We
-    look at consecutive tool_call items and stop at the first non-match.
-    """
-    out: list[dict[str, Any]] = []
-    if start == 0:
-        return out
-    prev = items[start - 1]
-    if prev.kind is not ItemKind.agent_msg:
-        return out
-    expected_idx = prev.metadata.get("assistant_idx") if prev.metadata else None
-    cursor = start
-    while cursor < len(items):
-        candidate = items[cursor]
-        if candidate.kind is not ItemKind.tool_call:
-            break
-        # Only consume tool_calls that came from this assistant turn.
-        cand_idx = candidate.metadata.get("assistant_idx") if candidate.metadata else None
-        if cand_idx != expected_idx:
-            break
-        out.append({"payload": _tool_call_payload(candidate)})
-        cursor += 1
-    return out
-
-
-def _tool_call_payload(item: ContextItem) -> dict[str, Any]:
-    """Build an OpenAI ``tool_calls[]`` entry from a stored tool_call item."""
-    tool_call_id = item.metadata.get("tool_call_id") if item.metadata else None
-    if not tool_call_id:
-        # Fall back to stripping the prefix; rejects items that weren't
-        # produced by this adapter.
-        tool_call_id = strip_prefix(item.id, _PREFIX_TOOL_CALL)
-    if not tool_call_id:
-        raise CatalogError(
-            f"tool_call item {item.id!r} cannot round-trip: missing "
-            "tool_call_id metadata and id is not openai-prefixed"
-        )
-    fn_name = item.metadata.get("function_name", "") if item.metadata else ""
-    arguments = item.metadata.get("arguments") if item.metadata else None
-    if arguments is None:
-        arguments = item.text
-    return {
-        "id": tool_call_id,
-        "type": item.metadata.get("tool_call_type", "function") if item.metadata else "function",
-        "function": {"name": fn_name, "arguments": arguments},
-    }
-
-
-# --- Shared helpers ---
-
-
-def _restore_assistant_content(item: ContextItem, meta: dict[str, Any]) -> Any:  # noqa: ANN401 — provider-shaped JSON
-    """Reconstruct assistant ``content``: None / list / string (PR #230)."""
-    if meta.get("content_is_null"):
-        return None
-    original = meta.get("original_content")
-    if isinstance(original, list):
-        return original
-    return item.text
-
-
-def _string_content(msg: dict[str, Any], idx: int, role: str) -> str:
-    """Coerce ``content`` to a string; collapse list-of-parts to joined text.
-
-    The list path supports OpenAI's multimodal `content` shape. The original
-    list payload is preserved separately in metadata so the round-trip is
-    lossless (see :func:`_from_assistant`).
-    """
-    content = msg.get("content")
-    if content is None:
-        return ""
-    if isinstance(content, str):
-        return content
-    if isinstance(content, list):
-        parts = [
-            str(p.get("text", ""))
-            if isinstance(p, dict) and p.get("type") == "text"
-            else json.dumps(p, sort_keys=True)
-            for p in content
-        ]
-        return "\n".join(parts)
-    raise CatalogError(
-        f"OpenAI {role} message at index {idx} has unsupported content type: "
-        f"{type(content).__name__}"
-    )
